i found it

+++ rust/src/mmu/mbc3.rs
@@ -6,7 +6,7 @@ impl MMU {
     pub(super) fn mbc3_read_rom(&self, loc: u16) -> u8 {
         match loc {
             ..0x4000 => self.rom[loc as usize],
-            loc => self.rom[(loc + 0x4000 * (self.rom_bank as u16 - 1)) as usize],
+            loc => self.rom[(loc as u32 + 0x4000 * (self.rom_bank as u32 - 1)) as usize],
         }
     }
